+++ inst/tinytest/testImport.R
@@ -1,0 +1,36 @@
+# Test importStudy()
+
+# Setup ------------------------------------------------------------------------
+
+source("tinytestSettings.R")
+using(ttdo)
+
+library(OmicNavigator)
+
+testStudyName <- "ABC"
+testStudyObj <- OmicNavigator:::testStudy(name = testStudyName)
+testStudyObj <- addPlots(testStudyObj, OmicNavigator:::testPlots())
+testModelName <- names(testStudyObj[["models"]])[1]
+testTestName <- names(testStudyObj[["tests"]][[1]])[1]
+testAnnotationName <- names(testStudyObj[["annotations"]])[1]
+
+minimalStudyObj <- OmicNavigator:::testStudyMinimal()
+minimalStudyName <- minimalStudyObj[["name"]]
+
+tmplib <- tempfile()
+dir.create(tmplib)
+libOrig <- .libPaths()
+.libPaths(c(tmplib, libOrig))
+suppressMessages(installStudy(testStudyObj))
+suppressMessages(installStudy(minimalStudyObj))
+
+# importStudy ------------------------------------------------------------------
+
+importStudy(testStudyName)
+
+importStudy(minimalStudyName)
+
+# Teardown ---------------------------------------------------------------------
+
+unlink(tmplib, recursive = TRUE, force = TRUE)
+.libPaths(libOrig)